backend/requirements.txt
selenium
chromedriver_autoinstaller
python-dotenv